[FIN-180] refactor: Container-Presenter 패턴 전체 적용
