--- app/components/SidePanel.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import Tabs from '@/app/components/Tabs';
-
-// TODO: 리소스 경로 관리
-export default function SidePanel() {
-  // const [isOpen, setIsOpen] = useState(true);
-
-  return (
-    <div>
-      <Tabs></Tabs>
-    </div>
-  );
-}
