+++ app/components/SmartDrag/DragRelatedArticles/DragRelatedArticlesContainer.tsx
@@ -1,0 +1,5 @@
+import DragRelatedArticlesView from '@/app/components/SmartDrag/DragRelatedArticles/DragRelatedArticlesView';
+
+export default function DragRelatedArticlesContainer() {
+  return <DragRelatedArticlesView />;
+}
