+++ app/components/SmartDrag/DragRelatedArticles/DragRelatedArticlesView.tsx
@@ -1,0 +1,5 @@
+import IntroView from '@/app/components/SmartDrag/Intro/IntroView';
+
+export default function DragRelatedArticlesView() {
+  return <div>관련 아티클 UI </div>;
+}
