+++ app/components/SmartDrag/Intro/IntroContainer.tsx
@@ -1,0 +1,5 @@
+import IntroView from '@/app/components/SmartDrag/Intro/IntroView';
+
+export default function IntroContainer() {
+  return <IntroView />;
+}
