+++ app/components/SmartDrag/SidePanel/SidePanelContainer.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import { useTooltipStore, useSidePanelStore } from '@/store/sidePanel';
+import SidePanelView from '@/app/components/SmartDrag/SidePanel/SidePanelView';
+
+export default function SidePanelContainer() {
+  const { isOpenSidePanel, setIsOpenSidePanel } = useSidePanelStore();
+  const { selectedMode } = useTooltipStore();
+
+  return (
+    <SidePanelView
+      selectedMode={selectedMode}
+      isOpenSidePanel={isOpenSidePanel}
+      setIsOpenSidePanel={setIsOpenSidePanel}
+    />
+  );
+}
